Fix sign extension when decoding Backward Size
The left shift was done as int, and the result was widened to long.
If the most significant byte of Backward Size had the highest bit set,
the widening sign-extended it, and the final backwardSize became
a non-positive value. That value slipped past the backwardSize >= pos
check in SeekableXZInputStream and lead to an out-of-range seek.

Fix by casting to long before shifting.

Fixes: https://github.com/tukaani-project/xz-java/pull/25

+++ src/org/tukaani/xz/common/DecoderUtil.java
@@ -66,7 +66,7 @@ public class DecoderUtil extends Util {
 
         streamFlags.backwardSize = 0;
         for (int i = 0; i < 4; ++i)
-            streamFlags.backwardSize |= (buf[i + 4] & 0xFF) << (i * 8);
+            streamFlags.backwardSize |= (long)(buf[i + 4] & 0xFF) << (i * 8);
 
         streamFlags.backwardSize = (streamFlags.backwardSize + 1) * 4;
